Dev code regions

+++ Assets/Scripts/BasicSMExtensions/States/RunSMState.cs
@@ -3,24 +3,55 @@ using BasicSM;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 
-namespace Utility.BasicFSMExtensions
+namespace BasicSMExtensions.States
 {
+    /// <summary>
+    /// State that loads and runs a sub-state machine from an Addressable GameObject asset.
+    /// </summary>
     public class RunSMState : StateComponentBase
     {
-        [SerializeField]private AssetReference _smAsset;
+        #region Serialized Parameters
+
+        /// <summary>
+        /// The addressable reference to the sub-state machine prefab.
+        /// </summary>
+        [SerializeField]
+        private AssetReference _smAsset;
+
+        #endregion
+
+        #region Non Serialized Parameters
+
+        /// <summary>
+        /// The instance of the sub-state machine that was spawned.
+        /// </summary>
         private StateMachineComponentBase subStateMachineInstance = null;
 
+        #endregion
+
+        #region Public Methods
+
+        /// <summary>
+        /// Loads and instantiates the sub-state machine, and initializes it.
+        /// </summary>
+        /// <param name="stateMachine">The parent state machine.</param>
+        /// <returns>Coroutine enumerator.</returns>
         public override IEnumerator OnEnter(IStateMachine stateMachine)
         {
             var loadSMAsset = Addressables.LoadAssetAsync<GameObject>(_smAsset);
             yield return loadSMAsset;
 
-            GameObject smInstance = Instantiate(loadSMAsset.Result, transform,false);
-            StateMachineComponentBase stateMachineComponentBase = smInstance.GetComponent<StateMachineComponentBase>();
-            
-            yield return stateMachineComponentBase.Initialize(stateMachine);
+            GameObject smInstance = Object.Instantiate(loadSMAsset.Result, transform, false);
+            subStateMachineInstance = smInstance.GetComponent<StateMachineComponentBase>();
+
+            yield return subStateMachineInstance.Initialize(stateMachine);
         }
 
+        /// <summary>
+        /// Delegates the OnUpdate loop to the sub-state machine instance.
+        /// </summary>
+        /// <param name="stateMachine">The parent state machine.</param>
+        /// <returns>Coroutine enumerator.</returns>
         public override IEnumerator OnUpdate(IStateMachine stateMachine)
         {
             while (subStateMachineInstance == null)
@@ -29,9 +60,13 @@ namespace Utility.BasicFSMExtensions
             }
 
             yield return subStateMachineInstance.OnUpdate();
-            
         }
 
+        /// <summary>
+        /// Delegates the OnFixedUpdate loop to the sub-state machine instance.
+        /// </summary>
+        /// <param name="stateMachine">The parent state machine.</param>
+        /// <returns>Coroutine enumerator.</returns>
         public override IEnumerator OnFixedUpdate(IStateMachine stateMachine)
         {
             while (subStateMachineInstance == null)
@@ -42,11 +77,21 @@ namespace Utility.BasicFSMExtensions
             yield return subStateMachineInstance.OnFixedUpdate();
         }
 
+        /// <summary>
+        /// Terminates and destroys the sub-state machine instance when exiting this state.
+        /// </summary>
+        /// <param name="stateMachine">The parent state machine.</param>
+        /// <returns>Coroutine enumerator.</returns>
         public override IEnumerator OnExit(IStateMachine stateMachine)
         {
-            yield return subStateMachineInstance.Terminate();
-            Destroy(subStateMachineInstance.gameObject);
-            subStateMachineInstance = null;
+            if (subStateMachineInstance != null)
+            {
+                yield return subStateMachineInstance.Terminate();
+                Object.Destroy(subStateMachineInstance.gameObject);
+                subStateMachineInstance = null;
+            }
         }
+
+        #endregion
     }
 }
